+++ include/databento/live_blocking.hpp
@@ -50,6 +50,8 @@ class LiveBlocking {
                  SType stype_in, UnixNanos start);
   void Subscribe(const std::vector<std::string>& symbols, Schema schema,
                  SType stype_in, const std::string& start);
+  void Subscribe(const std::vector<std::string>& symbols, Schema schema,
+                 SType stype_in, bool use_snapshot);
   // Notifies the gateway to start sending messages for all subscriptions.
   //
   // This method should only be called once per instance.
@@ -79,7 +81,7 @@ class LiveBlocking {
   std::string EncodeAuthReq(const std::string& auth);
   std::uint64_t DecodeAuthResp();
   void Subscribe(const std::string& sub_msg,
-                 const std::vector<std::string>& symbols);
+                 const std::vector<std::string>& symbols, bool use_snapshot);
   detail::TcpClient::Result FillBuffer(std::chrono::milliseconds timeout);
   RecordHeader* BufferRecordHeader();