+++ src/live_threaded.cpp
@@ -102,6 +102,11 @@ void LiveThreaded::Subscribe(const std::vector<std::string>& symbols,
   impl_->blocking.Subscribe(symbols, schema, stype_in, start);
 }
 
+void LiveThreaded::Subscribe(const std::vector<std::string>& symbols,
+                             Schema schema, SType stype_in, bool use_snapshot) {
+  impl_->blocking.Subscribe(symbols, schema, stype_in, use_snapshot);
+}
+
 void LiveThreaded::Start(RecordCallback callback) {
   Start({}, std::move(callback), {});
 }